+++ data/src/main/java/org/keelfy/eljur/data/repository/DepartmentRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 
 /**
- * @author Egor Kuzmin
+ * @author Yegor Kuzmin (keelfy)
  */
-public interface DepartmentRepository extends JpaRepository<Department, BigInteger> {
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
+
 }
